Fixes: Fixed a bug where the user would not have a profile picture if they changed the pic to the same format as the current one.
- changes: 
added styling to the profile.

+++ PresentationLayer/Scripts/upload.php
@@ -47,7 +47,8 @@ if ($uploadOk == 0) {
   if (move_uploaded_file($_FILES["fileToUpload"]["tmp_name"], $target_file)) {
     echo "The file ". htmlspecialchars( basename( $_FILES["fileToUpload"]["name"])). " has been uploaded.";
     SetProfilePicture($_SESSION['user'], $_FILES["fileToUpload"]["name"]);
-    unlink($oldPicture);
+    if($oldPicture!= $target_file)
+      unlink($oldPicture);
   } else {
     echo "Sorry, there was an error uploading your file.";
   }
